Return large device addresses as 64 bit integer

+++ src/Systems/Peer.cpp
@@ -2481,7 +2481,7 @@ PVariable Peer::getDeviceDescription(PRpcClientInfo clientInfo, int32_t channel,
       if (!_ip.empty() && (fields.empty() || fields.find("IP_ADDRESS") != fields.end())) description->structValue->insert(StructElement("IP_ADDRESS", PVariable(new Variable(_ip))));
 
       if (fields.empty() || fields.find("PHYSICAL_ADDRESS") != fields.end()) {
-        if (_address > 2147483647) description->structValue->emplace("PHYSICAL_ADDRESS", std::make_shared<BaseLib::Variable>((int64_t)_address));
+        if (_address > 2147483647) description->structValue->emplace("PHYSICAL_ADDRESS", std::make_shared<BaseLib::Variable>((int64_t)(uint32_t)_address));
         else description->structValue->emplace("PHYSICAL_ADDRESS", std::make_shared<BaseLib::Variable>(_address));
       }
